requirements.txt
reflex==0.8.26
python-dotenv>=1.0.0
openai>=1.0.0
psycopg2-binary>=2.9.0
